fix(accordion): rewrite logic
