+++ projects/angular/src/accordion/stepper/stepper-panel.html
@@ -1,6 +1,6 @@
 <ng-container *ngIf="panel | async; let panel">
   <div [ngClass]="getPanelStateClasses(panel)">
-    <div class="clr-accordion-header clr-stepper-header">
+    <div class="clr-accordion-header clr-stepper-header" role="heading" [attr.aria-level]="ariaLevel">
       <button
         type="button"
         class="clr-accordion-header-button"
