+++ projects/angular/src/accordion/stepper/stepper-panel.ts
@@ -15,6 +15,7 @@ import {
   OnInit,
   Optional,
   PLATFORM_ID,
+  SkipSelf,
   ViewChild,
 } from '@angular/core';
 import { FormGroupName, NgModelGroup } from '@angular/forms';
@@ -45,6 +46,9 @@ export class ClrStepperPanel extends ClrAccordionPanel implements OnInit {
   private subscriptions: Subscription[] = [];
 
   constructor(
+    @Optional()
+    @SkipSelf()
+    public override parent: ClrAccordionPanel,
     @Inject(PLATFORM_ID) private platformId: any,
     public override commonStrings: ClrCommonStringsService,
     @Optional() private formGroupName: FormGroupName,
@@ -53,7 +57,7 @@ export class ClrStepperPanel extends ClrAccordionPanel implements OnInit {
     ifExpandService: IfExpandService,
     cdr: ChangeDetectorRef
   ) {
-    super(commonStrings, stepperService, ifExpandService, cdr);
+    super(parent, commonStrings, stepperService, ifExpandService, cdr);
   }
 
   override get id(): string {
